refactor: Refactored layout structure - Added Layout folder and component - Changed styles
