+++ src/components/Layout/Layout.module.scss
@@ -1,0 +1,9 @@
+.layout {
+  overflow-x: hidden;
+  overflow-y: auto;
+}
+
+.content {
+  min-height: 100vh;
+  height: 100%;
+}
