+++ src/styles/base/base.scss
@@ -4,7 +4,6 @@
 :root {
   display: flex;
   flex-direction: column;
-  min-height: 100vh;
 }
 
 html {
